flake.nix: use git submodules instead of flake input for libshv
It is really easy to encounter broken state because nix's flake.lock is
not updated by the upstream developers and thus we end up with broker
build due to mismatching libshv and the shvspy code.

This pivots back to use submodules as upstream is using them to pin
their dependencies.

The github actions workaround is required to make it work with new Nix
versions.